- Moved `Input` component to screens/components as it can only be used by `Todo` screen.  - Deleted `Spinner` component as it's not in use.  - Added routes to a new view called `BookDetails`.  - Now, if you press a book, you can see it's details.
