+++ src/screens/BookDetails/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+
+const BookDetails = () => {
+  return (
+    <View>
+      <Text>Hola</Text>
+    </View>
+  );
+};
+
+export default BookDetails;
