+++ src/screens/Books/components/ListItem/index.js
@@ -1,24 +1,25 @@
 import React from 'react';
-import { Text, View, Image } from 'react-native';
+import { Text, View, Image, TouchableOpacity } from 'react-native';
 import PropTypes from 'prop-types';
 
 import CardSection from '../../../../components/CardSection';
 
 import styles from './styles';
 
-const ListItem = ({ elem }) => {
+const ListItem = ({ elem, onPress }) => {
   const { title, author, imageLink } = elem;
-
   return (
     <View>
       <CardSection>
-        <View style={styles.thumbnailContainer}>
-          <Image style={styles.thumbnail} source={{ uri: imageLink }} />
-        </View>
-        <View style={styles.headerContent}>
-          <Text style={styles.headerText}>{title}</Text>
-          <Text>{author}</Text>
-        </View>
+        <TouchableOpacity onPress={onPress}>
+          <View style={styles.thumbnailContainer}>
+            <Image style={styles.thumbnail} source={{ uri: imageLink }} />
+          </View>
+          <View style={styles.headerContent}>
+            <Text style={styles.headerText}>{title}</Text>
+            <Text>{author}</Text>
+          </View>
+        </TouchableOpacity>
       </CardSection>
     </View>
   );
