+++ src/screens/Todo/components/Input/index.js
@@ -3,9 +3,9 @@ import { TextInput, Keyboard } from 'react-native';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
-import { actionCreators as todoActions } from '../../redux/todo/actions';
-import IconButton from '../IconButton';
-import CardSection from '../CardSection';
+import { actionCreators as todoActions } from '../../../../redux/todo/actions';
+import IconButton from '../../../../components/IconButton';
+import CardSection from '../../../../components/CardSection';
 
 import styles from './styles';
